requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
crispy-bootstrap4==2025.6
cryptography==45.0.4
cssselect2==0.8.0
dj-database-url==3.0.0
Django==5.2.3
django-allauth==65.9.0
django-crispy-forms==2.4
django-formtools==2.5.1
django-js-asset==3.1.2
django-widget-tweaks==1.5.0
fonttools==4.58.4
gunicorn==23.0.0
idna==2.6
ollama>=0.1.0  # If using pip-based wrapper
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
PyPDF2==3.0.1
pyphen==0.17.2
pytest-django==4.5.2
python-docx==1.1.0
python-dotenv==1.1.1
python-magic-bin==0.4.14
requests==2.31.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.18
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
